move pkg/cluster/internal/patch to pkg/internal/patch
for use implementing containerd config patching at node image build time
