+++ pkg/internal/patch/resource.go
